Changes to header spacing and links
It doesn’t make sense to separate Offices from Services in the global
nav.  I know Jason S. typed it up that way, but that wasn’t ever our
intent; O&S is a global category that applies to a wide variety of
sites. No need to sweat over which is which.   This commit makes them
into a single list, in four columns, with a link to more.  (We feel
that this list could comfortably grow, even double, in length if people
want to fight about being there.)

Also changed the Schools nav to be 7/5 rather than 6/6, so the colleges
will form a tighter set.