#12 Institutional navigation drop down on site-specific nav pages
This works.  See here:
http://scu.babywhale.net/site-specific-search.php

It only pushes the nav down when you’re scrolled down the page, as
you’d expect.  And it won’t work to move it gradually; moving page
elements on hover will make you woozy. But the toggled position of the
top nav seems to work.    Please review!