Large commit for new site headers
This commit provides new header images and a more consistent approach
to coding header includes for the entire site.  Rules for use are
covered in index.php or at http://scu.babywhale.net.

Lots to discuss here, but the HTML/CSS approach hasn’t changed.

Note that we also removed the unnecessary /bootstrap-3.2.0/ directory,
and those changes are being pushed with this commit as well, making it
look even bigger than it actually is.