Added elements to gallery module
Added a header element and a “View All” link to the gallery module.
Results can be seen in the Thriving Neighbors sample page, to be
uploaded later today.

The markup for a gallery, including these new elements, is committed in
a separate file (in a “module markup” folder where we can put more code
snippets if it works for @bwashburnscu and team).

(Note that the commit here is LESS:  compiled CSS has already been
pushed in the last commit.)