Adjusts spacing for templates and layouts (revised demos)
This commit removes padding from the #content element and replaces it with margins on the appropriate elements.

Pages that use the "core" heading will margin adjustments to clear the absolutely-positioned overhanging menu. For other pages, they will most likely have an mb-4 between each of the major blocking elements (this matches the previous design).